+++ app/components/TimelineStreamingButton.js
@@ -4,8 +4,7 @@ import { FontAwesome } from "@expo/vector-icons";
 import { useSelector, useDispatch } from "react-redux";
 
 import { ThemeContext } from "react-native-elements";
-import { start, stop, receive } from "../actions/actioncreators/streaming";
-import { STREAMING } from "../constants/api";
+import { on as streamingOn, off as streamingOff } from "../util/stream";
 
 const CurrentUserReducerSelector = state => state.currentUserReducer;
 
@@ -21,57 +20,20 @@ export default function TimelineStreamingButton({ type }){
     useEffect(() => {
         if (enabled && instance){
             // ON
-            let stream;
-            switch (type) {
-                case "federal":
-                    stream = "public";
-                    break;
-                case "local":
-                    stream = "public:local";
-                    break;
-                case "home":
-                default:
-                    stream = "user";
-                    break;
-            }
-            let url = instance.urls.streaming_api + STREAMING.url + "?access_token=" + access_token + "&stream=" + stream;
-            webSocket.current = new WebSocket(url);
-            webSocket.current.onopen = () => {
-                // connection opened
-                console.log("[WS] OPEN:" + type);
-            };
-            webSocket.current.onmessage = (message) => {
-                let data = JSON.parse(message.data);
-                if (data !== null){
-                    let { event, payload } = data;
-                    dispatch(receive(type, event, JSON.parse(payload)));
-                }
-            };
-            webSocket.current.onclose = (event) => {
-                console.log("[WS] CLOSE:" + type + " event:" + JSON.stringify(event));
-            };
-            webSocket.current.onerror = (event) => {
-                console.log("[WS] ERROR:" + type + " event:" + JSON.stringify(event));
-                dispatch(stop(type));
-                useEnabled(false);
-            };
-            dispatch(start(type));
+            streamingOn(webSocket, dispatch, useEnabled, type, instance, access_token);
+            console.log("[WS] ON:" + type);
         } else {
-            dispatch(stop(type));
-            if (webSocket.current !== null && webSocket.current.readyState === WebSocket.OPEN){
-                webSocket.current.close();
-            }
+            // OFF
+            streamingOff(webSocket, dispatch, useEnabled, type);
+            console.log("[WS] OFF:" + type);
         }
         return;
     }, [enabled]);
     useEffect(() => {
         return () => {
-            // unmount
-            dispatch(stop(type));
-            if (webSocket.current !== null && webSocket.current.readyState === WebSocket.OPEN){
-                webSocket.current.close();
-                console.log("[WS] UNMOUNTCLOSE:" + type);
-            }
+            // UNMOUNT
+            streamingOff(webSocket, dispatch, useEnabled, type);
+            console.log("[WS] UNMOUNTCLOSE:" + type);
         };
     }, []);
     return (
